refactor: address PR review feedback
- Preserve server error response in HTTPError (response=resp)
- Add MAX_UPLOAD_FILE_SIZE (100MB) guard in _upload
- Use UTC timezone for request_time
- Distinguish None vs empty key in _init error messages
- Remove 180s timeout from init (only needed for eval)
- Rename Client.file param to file_path for consistency
- Fix Tuple[int, int] to Tuple[float, float] in api.py
- Add retry safety comment for bytes upload
- Add tests for file size guard and empty key

+++ podonos/core/api.py
@@ -181,7 +181,7 @@ class APIClient:
         endpoint: str,
         data: Dict[str, Any],
         headers: Optional[Dict[str, str]] = None,
-        timeout: Tuple[int, int] = (5, 30),
+        timeout: Tuple[float, float] = (5, 30),
     ) -> Response:
         request_header = self._headers if headers is None else headers
